requirements.txt
absl-py==0.7.1
astor==0.8.0
Click==7.0
decorator==4.4.0
Flask==1.1.1
Flask-Cors==3.0.8
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.17.0
protobuf==3.9.0
pymongo==2.8
python-dotenv==0.10.3
PyYAML==5.1.1
scipy==1.3.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
uWSGI==2.0.18
validators==0.13.0
Werkzeug==0.15.5
wrapt==1.11.2
